Move OAuthServerExceptionTest to appropriate folder
